pyproject.toml
[project]
name = "pyhctsa"
version = "0.1.0"
description = ""
authors = [
    {name = "Joshua Moore",email = "jmoo2880@sydney.edu.au"}
]
readme = "README.md"
requires-python = "^3.10"
dependencies = [
    "jpype1 @ git+https://github.com/jpype-project/jpype.git@v1.5.2",
    "scipy (>=1.15.1,<2.0.0)",
    "numpy (==2.1)",
    "scikit-learn (>=1.6.1,<2.0.0)",
    "statsmodels (>=0.14.4,<0.15.0)",
    "pywavelets (>=1.8.0,<2.0.0)",
    "numba (>=0.61.0,<0.62.0)",
    "antropy (>=0.1.8,<0.2.0)",
    "nolitsa @ git+https://github.com/manu-mannattil/nolitsa.git",
    "loguru (>=0.7.3,<0.8.0)",
    "lmfit (>=1.3.2,<2.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
